Thonny (#6)
Include a pre-downloaded Thonny package for installation where useful.

Thonny's installer doesn't provide an interface that we can use to properly install it during the creation of the Live CD image. We could contribute to Thonny to implement this, but for now let's be Lean and try a low-tech approach first.