Clearly show build details (#12)
Often I find myself wondering what version of the OS a laptop is running. Should it be updated?

This PR prints build information on the desktop wallpaper, so that we quickly identify what version of the software the laptop is carrying. It includes the date of the build and the Git branch name (which should normally be `master`, but sometimes won't).

![w2c3-fingerprint](https://user-images.githubusercontent.com/36066/52180819-c2743a80-27e2-11e9-9b79-b5eed1f5de2c.png)