Mirror screen shortcut (#13)
At show&tell time, getting the big screen to work is a lot of faff, involving a
bunch of menus and clicks for each individual student. And the result is not
even great: it normally creates virtual screen with the resolution of the large
screen, and leaves the small screen covering only a portion of this larger
screen. As a result, the presenter's screen only shows a portion of the whole
desktop. It also continues after the secondary screen is unplugged (although we
are less worried about this because the student should return the laptop at
this point).

This PR introduces a desktop shortcut that uses
[XRandR](https://www.x.org/wiki/Projects/XRandR/) to figure out the resolutions
of all screens and scale things so that the big screen is a mirror of the
presenter's laptop.